Suggest ratio checks

Suggest ratio checks

write_ratio_check(d, vars = names(d), file = stdout(), lin_cor = 0.95)

suggest_ratio_check(d, vars = names(d), lin_cor = 0.95)

Arguments

d
data.frame, used to generate the checks

vars
character optionally the subset of variables to be used.

file
file to which the checks will be written to.

lin_cor
threshold for abs correlation to be included (details)

Examples

data(SBS2000, package="validate")

# generates upper and lower checks for the
# ratio of two variables if their correlation is
# bigger then `lin_cor`
suggest_ratio_check(SBS2000, lin_cor=0.98)

#> Object of class 'validator' with 10 elements:
#>  RC1 : turnover >= 0 * total.rev
#>  RC2 : turnover <= 9.07 * total.rev
#>  RC3 : other.rev >= -0.1 * staff.costs
#>  RC4 : other.rev <= 34.55 * staff.costs
#>  RC5 : other.rev >= -0.01 * total.costs
#>  RC6 : other.rev <= 1.27 * total.costs
#>  RC7 : staff.costs >= 0 * total.costs
#>  RC8 : staff.costs <= 0.99 * total.costs
#>  RC9 : other.rev >= -2.8 * profit
#>  RC10: other.rev <= 4.72 * profit